pyproject.toml
[project]
name = "post-train"
version = "0.1.0"
description = "RL training system for coding agents using GRPO"
requires-python = ">=3.10"
dependencies = [
    "trl>=0.23.0",
    "transformers>=4.48.0",
    "accelerate>=1.3.0",
    "peft>=0.14.0",
    "datasets>=3.2.0",
    "vllm>=0.8.5",
    "torch>=2.5.0",
    "docker>=7.0.0",
    "wandb>=0.19.0",
    "evalplus>=0.3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
